[luci] Remove duplicate codes for quantization
This removes duplicate codes for quantization

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
